CM-851: Moving the demo classes to separate files and adding error messages in comments
