+++ csharp-intermediate-topics/UsingStaticAnonymousFunctionsInCSharp/UsingStaticAnonymousFunctionsInCSharp/Demo/DemoStaticBase.cs
@@ -1,0 +1,7 @@
+﻿namespace UsingStaticAnonymousFunctionsInCSharp.Demo;
+
+// base class for DemoStaticDerivative
+public class DemoStaticBase
+{
+    public double numberInBase = 3;
+}
